-Refactored a bit
